change parameter from int to long in interface

+++ Door2Door-API/Models/Interfaces/IEntityRepository.cs
@@ -3,5 +3,5 @@
 public interface IEntityRepository<T> where T : class 
 {
     Task<IEnumerable<T>> GetAllAsync();
-    Task<T?> GetByIdAsync(int id);
+    Task<T?> GetByIdAsync(long id);
 }
